integer attention mask

+++ ldm/modules/encoders/modules.py
@@ -162,7 +162,7 @@ class FrozenCLIPEmbedder(AbstractEncoder):
         batch_encoding = self.tokenizer(text, truncation=True, max_length=self.max_length, return_length=True,
                                         return_overflowing_tokens=False, padding="max_length", return_tensors="pt")
         tokens = batch_encoding["input_ids"].to(self.device)
-        outputs = self.transformer(input_ids=tokens, attention_mask=tokens != self.tokenizer.pad_token_id)
+        outputs = self.transformer(input_ids=tokens, attention_mask=(tokens != self.tokenizer.pad_token_id).long())
 
         z = outputs.last_hidden_state
         return z
